fix warning
git-svn-id: https://yuma.svn.sourceforge.net/svnroot/yuma/trunk@1259 1ffd1d58-2efc-45df-af60-190a33ec9863

+++ netconf/src/ncx/ses.c
@@ -1906,7 +1906,7 @@ status_t
             }
 
             if (res != NO_ERR || 
-                ret < scb->readbuffsize || 
+                ((uint32)ret < scb->readbuffsize) || 
                 scb->rdfn == NULL) {
 #ifdef SES_DEBUG_TRACE
                 if (LOGDEBUG3) {
